/webpack

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video"
+import User from "../models/User";
 
 /*Video.find({}, (error, videos) => {
     if(error){
@@ -11,41 +12,74 @@ import Video from "../models/Video"
 */
 
 export const home = async (req, res) => {
-    const videos = await Video.find({}).sort({createdAt: "1"});
-    return res.render("home", {pageTitle : "Home", videos });
+    const videos = await Video
+        .find({})
+        .sort({createdAt: "1"});
+    return res.render("home", {
+        pageTitle: "Home",
+        videos
+    });
 };
 
 export const watch = async (req, res) => {
-    const { id } = req.params;
-    const video = await Video.findById(id);
-    if( !video) {
-    return res.render("404", {pageTitle:"video not found"});
+    const {id} = req.params;
+    const video = await Video
+        .findById(id)
+        .populate("owner");
+    console.log(video);
+    if (!video) {
+        return res.render("404", {pageTitle: "video not found"});
 
     }
-    return res.render("watch", { pageTitle: video.title, video });
+    return res.render("watch", {
+        pageTitle: video.title,
+        video
+    });
 };
 
 export const getEdit = async (req, res) => {
-    const { id } = req.params;
+    const {id} = req.params;
+    const {user: {
+            _id
+        }} = req.session;
     const video = await Video.findById(id);
-    if( !video) {
-        return res.status(404).render("404", {pageTitle:"video not found"});
-    
-        }
-    return res.render("edit", {pageTitle:`Edit:${video.title}`, video});
+    if (!video) {
+        return res
+            .status(404)
+            .render("404", {pageTitle: "video not found"});
+    }
+    console.log(typeof video.owner, typeof _id);
+    if (String(video.owner) !== String(_id)) {
+        return res
+            .status(403)
+            .redirect("/");
+    }
+    return res.render("edit", {
+        pageTitle: `Edit:${video.title}`,
+        video
+    });
 };
 export const postEdit = async (req, res) => {
-    const { id } = req.params;
+    const {user: {
+            _id
+        }} = req.session;
+    const {id} = req.params;
     const {title, description, hashtags} = req.body;
-    const video = await Video.exists({ _id: id });
-    if( !video) {
-        return res.status(404).render("404", {pageTitle:"video not found"});
-        };
-        await Video.findByIdAndUpdate(id, {
-            title,
-            description,
-            hashtags: Video.formatHashtags(hashtags),
-        });
+    const video = await Video.exists({_id: id});
+    if (!video) {
+        return res
+            .status(404)
+            .render("404", {pageTitle: "video not found"});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res
+            .status(403)
+            .redirect("/");
+    }
+    await Video.findByIdAndUpdate(
+        id,
+        {title, description, hashtags: Video.formatHashtags(hashtags)}
+    );
     return res.redirect(`/videos/${id}`);
 };
 
@@ -54,40 +88,66 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
-    const { title, description, hashtags } = req.body;
-   try { 
-        await Video.create({
-        title,
-        description,
-        hashtags: Video.formatHashtags(hashtags),
-    });
-    return res.redirect("/");
-} catch (error) {
-    return res.render("upload", {
-        pageTitle: "Upload Video",
-        errorMessage: error._message,
-    });
-    
+    const {user: {
+            _id
+        }} = req.session;
+    const {path: fileUrl} = req.file
+    const {title, description, hashtags} = req.body;
+    try {
+        const newVideo = await Video.create(
+            {title, description, fileUrl, owner: _id, hashtags: Video.formatHashtags(hashtags)}
+        );
+        const user = await User.findById(_id);
+        user
+            .videos
+            .push(newVideo._id);
+        user.save();
+        return res.redirect("/");
+    } catch (error) {
+        console.log(error);
+        return res.render("upload", {
+            pageTitle: "Upload Video",
+            errorMessage: error._message
+        });
+
     }
 };
 
 export const deleteVideo = async (req, res) => {
-    const { id } = req.params;
+    const {id} = req.params;
+    const {user: {
+            _id
+        }} = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res
+            .status(404)
+            .render("404", {pageTitle: "video not found"});
+
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res
+            .status(403)
+            .redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 
 };
 
 export const search = async (req, res) => {
-    const { keyword } = req.query;
-    let videos = [] ;
+    const {keyword} = req.query;
+    let videos = [];
     if (keyword) {
         videos = await Video.find({
             title: {
                 $regex: new RegExp(keyword, "i")
-            }, 
+            }
         });
-    console.log(videos)
+        console.log(videos)
     }
-    return res.render("search", {pageTitle:"Search", videos} );
+    return res.render("search", {
+        pageTitle: "Search",
+        videos
+    });
 };
